requirements.txt
Flask==2.0.1
google-cloud-vision==2.4.2
langchain==0.0.1
faiss-cpu==1.7.1
openai==0.10.2
gunicorn==20.1.0
